Setup manual dependency injection, created another test and mock
