+++ index.js
@@ -7,13 +7,25 @@ const mongoose = require('mongoose');
 
 // MongoDB URL from the docker-compose file
 //const dbHost = 'mongodb://database/mean-docker';
-const dbHost = 'mongodb://database/mean-docker';
+
+//Use if running app without docker (after starting mongo manually)
+const dbHost = 'mongodb://localhost:27017/time-series';
 
 // Connect to mongodb
 mongoose.connect(dbHost);
 
-// Get our API routes
-const api = require('./app/routes/users');
+// Get our data layer
+const userModel = require('./app/models/userModel');
+
+//Get our business layer
+const usersService = require('./app/data-services/usersService')({
+    userModel: userModel
+})
+
+// Get our API layer
+const usersApi = require('./app/routes/usersApi')({
+    usersService: usersService
+});
 
 const app = express();
 
@@ -22,7 +34,7 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 
 // Set our api routes
-app.use('/', api);
+app.use('/', usersApi);
 
 /**
  * Get port from environment and store in Express.
